v0.6.1 - Parche: Arquitectura mejorada y configuración de ventana
NUEVAS CARACTERÍSTICAS:
- Nueva arquitectura mejorada para gestión de asociados
- Control de tamaño mínimo de ventana en aplicaciones desktop
- Estructura organizada por secciones y componentes reutilizables

REFACTORIZACIÓN:
- Separación de AsociadosMainView en secciones independientes
- Componentes de búsqueda, perfil y acciones modularizados
- Widgets compartidos para mejor reutilización
- Diálogos de confirmación y opciones centralizados

CORRECCIONES:
- Solucionado error de ScrollController en componentes con scroll
- Actualizado uso de .withOpacity() a .withValues(alpha:) para Flutter actual
- Corregidos imports y dependencias faltantes

DESKTOP:
- Configuración de window_manager para control de ventana
- Tamaño mínimo establecido y sin límite máximo
- Soporte para maximizar manteniendo límites sensatos

ESTRUCTURA:
- sections/search_section/ - Componentes de búsqueda
- sections/profile_section/ - Información del asociado
- sections/actions_section/ - Panel de acciones
- shared/widgets/ - Componentes reutilizables
- shared/dialogs/ - Diálogos modales

DEPENDENCIAS:
- Agregado window_manager ^0.3.7 para control de ventana desktop"
