+++ lib/config/window_config.dart
@@ -1,0 +1,35 @@
+import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart';
+import 'package:window_manager/window_manager.dart';
+
+class WindowConfig {
+  // Configuración de tamaños
+  static const double minWidth = 1500.0;   
+  static const double minHeight = 1000.0;   
+
+  // Inicializar ventana
+  static Future<void> initialize() async {
+    if (!kIsWeb && (defaultTargetPlatform == TargetPlatform.windows ||
+        defaultTargetPlatform == TargetPlatform.linux ||
+        defaultTargetPlatform == TargetPlatform.macOS)) {
+      
+      await windowManager.ensureInitialized();
+
+      const WindowOptions windowOptions = WindowOptions(
+        size: Size(minWidth, minHeight),
+        minimumSize: Size(minWidth, minHeight), 
+        center: true,
+        backgroundColor: Colors.white,
+        skipTaskbar: false,
+        titleBarStyle: TitleBarStyle.normal,
+      );
+
+      await windowManager.waitUntilReadyToShow(windowOptions, () async {
+        await windowManager.show();
+        await windowManager.focus();
+      });
+
+      await windowManager.setTitle('GestAsocia');
+    }
+  }
+}
