+++ lib/widgets/dashboard/modules/gestion_asociados/sections/profile_section/profile_section.dart
@@ -1,18 +1,31 @@
 import 'package:flutter/material.dart';
-import '../../../../../utils/app_theme.dart';
-import 'asociado_header_card.dart';
-import 'asociado_personal_info.dart';
-import 'asociado_family_charges.dart';
+import '../../../../../../utils/app_theme.dart';
+import 'components/profile_header.dart';
+import 'components/personal_info_card.dart';
+import 'components/family_charges_card.dart';
 
-class AsociadoProfileCard extends StatelessWidget {
+class ProfileSection extends StatefulWidget {
   final Map<String, dynamic> asociado;
   final VoidCallback onEdit;
 
-  const AsociadoProfileCard({
+  const ProfileSection({
     super.key,
     required this.asociado,
     required this.onEdit,
   });
+
+  @override
+  State<ProfileSection> createState() => _ProfileSectionState();
+}
+
+class _ProfileSectionState extends State<ProfileSection> {
+  final ScrollController _scrollController = ScrollController();
+
+  @override
+  void dispose() {
+    _scrollController.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
@@ -23,8 +36,8 @@ class AsociadoProfileCard extends StatelessWidget {
         boxShadow: [
           BoxShadow(
             color: Theme.of(context).brightness == Brightness.dark 
-                ? Colors.black.withOpacity(0.3)
-                : Colors.grey.withOpacity(0.1),
+                ? Colors.black.withValues(alpha: 0.3)
+                : Colors.grey.withValues(alpha: 0.1),
             blurRadius: 10,
             offset: const Offset(0, 4),
           ),
@@ -33,25 +46,27 @@ class AsociadoProfileCard extends StatelessWidget {
       child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
         children: [
-          // Header con foto y datos básicos
-          AsociadoHeaderCard(
-            asociado: asociado,
-            onEdit: onEdit,
+          // Header del perfil
+          ProfileHeader(
+            asociado: widget.asociado,
+            onEdit: widget.onEdit,
           ),
           
-          // Contenido scrolleable
+          // Contenido scrolleable con controller
           Expanded(
             child: Scrollbar(
+              controller: _scrollController,
               thumbVisibility: true,
               child: SingleChildScrollView(
+                controller: _scrollController,
                 child: Column(
                   crossAxisAlignment: CrossAxisAlignment.start,
                   children: [
                     // Información personal
-                    AsociadoPersonalInfo(asociado: asociado),
+                    PersonalInfoCard(asociado: widget.asociado),
                     
                     // Cargas familiares
-                    AsociadoFamilyCharges(asociado: asociado),
+                    FamilyChargesCard(asociado: widget.asociado),
                   ],
                 ),
               ),
