+++ pubspec.yaml
@@ -59,6 +59,9 @@ dependencies:
   # Utilidades adicionales (nuevas dependencias)
   permission_handler: ^11.3.1
 
+  # Gestión de ventana para aplicaciones de escritorio
+  window_manager: ^0.3.7
+
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
